Changed files for jombo23's mod, based on the most similar TPT++ commit found
Last modified file in source looks like src/simulation/elements/ALPH.cpp,
modified 11:17 Nov 18 2012.

Generated and object files exist in the original source with a
modification time of around 18:00 Nov 18 2012.

Therefore, this is probably the same source (or very similar) as was
used for V5, for which a Windows .exe was posted with a modification time
of 22:07 Nov 18 2012.